Code refactor
Reindented and restyled various parts of the code.

Removed dependency on chardet and requests module as they simple things
and were not really needed.

Merged GentooBot and GentooBotFrame into a single class as
GentooBotFrame was redundant and did nothing special.

Updated lastfm api keys.

Fixed a couple of small bugs.